JBEHAVE-1241 Separate given stories that are not meant to run directly or in a batch.
