+++ examples/core/src/main/java/org/jbehave/examples/core/stories/given/GivenStoriesLoadedByRelativePath.java
@@ -1,10 +1,10 @@
-package org.jbehave.examples.core.stories;
+package org.jbehave.examples.core.stories.given;
 
 import org.jbehave.core.configuration.Configuration;
 import org.jbehave.core.io.RelativePathCalculator;
 import org.jbehave.examples.core.CoreStory;
 
-public class GivenRelativePathStory extends CoreStory {
+public class GivenStoriesLoadedByRelativePath extends CoreStory {
 
     @Override
     public Configuration configuration() {
