+++ examples/core/src/main/java/org/jbehave/examples/core/stories/given/given_stories_loaded_by_relative_path.story
