+++ examples/core/src/main/java/org/jbehave/examples/core/stories/given/parametrised.story
