+++ examples/core/src/main/java/org/jbehave/examples/core/stories/given_stories_parametrised_by_anchor.story
@@ -4,7 +4,7 @@ The presence of the anchor implies that the scenario is executed normally and no
 Meta: @theme parametrisation
 @assetClass FX
 
-GivenStories: org/jbehave/examples/core/stories/parametrised.story#{0}
+GivenStories: org/jbehave/examples/core/stories/given/parametrised.story#{0}
               
 When the stock is traded at price 1.1
 Then the alert status is ON
