+++ examples/core/src/main/java/org/jbehave/examples/core/stories/given_stories_parametrised_by_meta.story
@@ -3,7 +3,7 @@ A story using GivenStories at story level parametrised by meta properties
 Meta: @theme parametrisation
 @assetClass FX @symbol STK1 @threshold 1.0
 
-GivenStories: org/jbehave/examples/core/stories/parametrised.story
+GivenStories: org/jbehave/examples/core/stories/given/parametrised.story
 
 Scenario: A scenario that is executed after the given stories parametrised by meta
