+++ examples/core/src/main/java/org/jbehave/examples/core/stories/skip_before_and_after_scenario_steps_if_given_story.story
@@ -1,6 +1,6 @@
 Scenario: A scenario where before and after steps are executed only once even if there is a given story
 
-GivenStories: org/jbehave/examples/core/stories/parametrised.story#{0}
+GivenStories: org/jbehave/examples/core/stories/given/parametrised.story#{0}
               
 When the stock is traded at price 1.1
 Then the alert status is ON
